Rename Pointed -> Pure; add Zero, One, and Extract.
This commit embraces Miles' position -- that type classes
like Pure or Empty are "incomplete" versions of a complete
type class (like Applicative or Monoid) and should be
named accordingly.

This commit also adds Zero and One (based on the additive
and multiplicative type classes from algebra) as well as
Extract (based on Comonad).